+++ Functions/LinearFunction.cs
@@ -7,19 +7,19 @@ using System.Threading.Tasks;
 namespace NeuralNetwork.Functions {
 	public class LinearFunction : GeneralFunction {
 
-		public LinearFunction(float slope) {
+		public LinearFunction(double slope) {
 			param = slope;
 		}
 
-		public override float Function(float x) {
+		public override double Function(double x) {
 			return x * param;
 		}
 
-		public override float Inverse(float y) {
+		public override double Inverse(double y) {
 			return y / param;
 		}
 
-		public override bool SetParameterFor(float x, float y) {
+		public override bool SetParameterFor(double x, double y) {
 			param = y / x;
 			return true;
 		}
